requirements.txt
pycryptodome>=3.19.0
requests>=2.31.0
adbutils>=1.2.6
